+++ NEWS.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.15.14] 27/08/2024
 
+### Added
+
+* A helpful error message when `ProductManifold` is used without `RecursiveArrayTools.jl`.
+
 ### Changed
 
 * `representation_size` for `ProductManifold` now returns `nothing` instead of a one-element tuple. This change makes it easier to notice errors caused by not having `RecursiveArrayTools.jl` loaded.